+++ lighthouse-plugin-publisher-ads/audits/ad-request-critical-path.js
@@ -17,9 +17,9 @@ const {auditNotApplicable} = require('../utils/builder');
 const {AUDITS, NOT_APPLICABLE} = require('../messages/messages');
 const {Audit} = require('lighthouse');
 const {formatMessage} = require('../messages/format');
+const {getAbbreviatedUrl, trimUrl} = require('../utils/resource-classification');
 const {getAdCriticalGraph} = require('../utils/graph');
 const {getTimingsByRecord} = require('../utils/network-timing');
-const {URL} = require('url');
 
 /** @typedef {LH.Artifacts.NetworkRequest} NetworkRequest */
 /** @typedef {LH.Gatherer.Simulation.NodeTiming} NodeTiming */
@@ -147,33 +147,6 @@ function computeDepth(requests) {
     }
   }
   return hops;
-}
-
-/**
- * Extracts the request from a URL.
- * @param {string} url
- * @return {string}
- */
-function getAbbreviatedUrl(url) {
-  const u = new URL(trimUrl(url));
-  const parts = u.pathname.split('/');
-  if (parts.length > 4) {
-    u.pathname = [...parts.splice(0, 4), '...'].join('/');
-  }
-  return u.toString();
-}
-
-/**
- * Removes the query string from the URL.
- * @param {string} url
- * @return {string}
- */
-function trimUrl(url) {
-  const u = new URL(url);
-  const PATH_MAX = 60;
-  const path = u.pathname.length > PATH_MAX ?
-    u.pathname.substr(0, PATH_MAX) + '...' : u.pathname;
-  return u.origin + path;
 }
 
 /**
